refactor portal resource search
