+++ apps/portals/ampals/src/pages/Search.tsx
@@ -1,19 +1,10 @@
-import { PortalSearchPage } from '@sage-bionetworks/synapse-portal-framework/components/PortalSearch/PortalSearchPage'
 import { portalSearchPageConfigs, searchPageTabs } from '@/config/searchConfig'
-import { useParams } from 'react-router'
+import ResourceSearchPage from '@sage-bionetworks/synapse-portal-framework/components/PortalSearch/ResourceSearchPage'
 
 function Search() {
-  const params = useParams()
-  const resourceType = params['resourceType']
-
-  const selectedTabIndex = searchPageTabs.findIndex(
-    tab => tab.path === resourceType,
-  )
-
   return (
-    <PortalSearchPage
-      selectedTabIndex={selectedTabIndex === -1 ? undefined : selectedTabIndex}
-      configs={portalSearchPageConfigs}
+    <ResourceSearchPage
+      portalSearchPageConfigs={portalSearchPageConfigs}
       searchPageTabs={searchPageTabs}
     />
   )
