+++ apps/portals/cancercomplexity/src/pages/Search.tsx
@@ -1,23 +1,14 @@
-import { PortalSearchPage } from '@sage-bionetworks/synapse-portal-framework/components/PortalSearch/PortalSearchPage'
 import {
   portalSearchPageConfigs,
   roleMapping,
   searchPageTabs,
 } from '@/config/searchConfig'
-import { useParams } from 'react-router'
+import ResourceSearchPage from '@sage-bionetworks/synapse-portal-framework/components/PortalSearch/ResourceSearchPage'
 
 function Search() {
-  const params = useParams()
-  const resourceType = params['resourceType']
-
-  const selectedTabIndex = searchPageTabs.findIndex(
-    tab => tab.path === resourceType,
-  )
-
   return (
-    <PortalSearchPage
-      selectedTabIndex={selectedTabIndex === -1 ? undefined : selectedTabIndex}
-      configs={portalSearchPageConfigs}
+    <ResourceSearchPage
+      portalSearchPageConfigs={portalSearchPageConfigs}
       searchPageTabs={searchPageTabs}
       roleMapping={roleMapping}
     />
